async load based on extent
