+++ app/src/components/map/LayerLoaderHelpers/FeatureCollectionToMarkerClusters.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+const FeatureCollectionToMarkerClusters: React.FC = (props) => {
+  return <></>;
+};
+export default FeatureCollectionToMarkerClusters;
